Move DOI deletion code to doi.py
